Added 'git ffb' - to fast forward a branch to whatever it is remotely
It's actually a force, so it's more destructive than you think. But...
well - it should only be done by a user who is conscious of the fact
that a branch they're not checked out on is behind its remote. So... If
you wanna rebase onto main but REMOTE main instead of LOCAL main, you
can just say git ffb main -> git rebase main. Instead of git rebase
origin/main (and maybe needing to git fetch, first)